requirements.txt
beautifulsoup4==4.11.1
blis==0.7.9
bs4==0.0.1
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
dateparser==1.1.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.5.0/en_core_web_trf-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
filelock==3.9.0
fonttools==4.38.0
huggingface-hub==0.12.0
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joypy==0.2.6
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.0
murmurhash==1.0.9
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pathy==0.10.1
patsy==0.5.3
Pillow==9.4.0
plotly==5.14.1
preshed==3.0.8
pydantic==1.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
ridgeplot==0.1.21
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.1
srsly==2.4.5
statsmodels==0.14.0
tenacity==8.2.2
thinc==8.1.7
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
typer==0.7.0
typing-extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.13
wasabi==1.1.1
wordcloud==1.8.2.2
zipp==3.14.0
